requirements.txt
gdeltPyR
google-cloud-bigquery
pandas
requests
pyarrow
db-dtypes
